Python/Tooling/LangChain/Tutorial/chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.3.1",
    "beautifulsoup4>=4.14.3",
    "bs4>=0.0.2",
    "cassio>=0.1.10",
    "chromadb>=1.4.0",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.127.1",
    "groq>=0.37.1",
    "langchain-community>=0.0.38",
    "langchain-core>=0.1.53",
    "langchain-google-genai>=1.0.4",
    "langchain-groq>=0.1.5",
    "langchain-objectbox>=0.1.0",
    "langchain-openai>=0.1.7",
    "langchainhub>=0.1.21",
    "langserve>=0.2.3",
    "pypdf>=6.5.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "sse-starlette>=3.1.1",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
    "wikipedia>=1.4.0",
]
